yks-assistant-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
Pillow
requests
together
python-dotenv
